Threads....

+++ phoned/notify.c
@@ -27,7 +27,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  */
-/* $Amigan: phoned/phoned/notify.c,v 1.1 2005/06/02 02:40:53 dcp1990 Exp $ */
+/* $Amigan: phoned/phoned/notify.c,v 1.2 2005/06/02 20:49:26 dcp1990 Exp $ */
 #include <fcntl.h>
 #include <ctype.h>
 #include <unistd.h>
@@ -82,7 +82,7 @@ int cid_notify(cid_t* c)
 	char* msg;
 	int s;
 	struct sockaddr_in sin;
-	short on = 0x1;
+	int on = 0x1;
 	addrsll_t *cur;
 	len = strlen(c->number) + strlen(c->name) + 8 + 5 + 4;
 	msg = malloc(len * sizeof(char));
